pyproject.toml
[tool.poetry]
name = "dapp"
version = "1.0.0"
description = ""
authors = [
    "Kwabena Aning <kwabena.aning@gmail.com>"
]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.1.2"
gunicorn = "^23.0.0"
python-dotenv= "^1.1.1"
solana = "^0.36.9"
pynacl = "^1.6.0"
pyjwt = "^2.10.1"
flask-limiter = "^4.0.0"
base58 = "^2.1.1"
mykobo-py = { git = "https://github.com/mykobo/mykobo-py.git", rev = "1.4.8" }
#mykobo-py = { path = "../mykobo-py", develop=true  }
flask-wtf = "^1.2.2"
pycountry = "^24.6.1"
email-validator = "^2.3.0"
schwifty = "^2025.9.0"
jinja-partials = "^0.3.0"
boto3 = "^1.40.55"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.1.0"
psycopg2-binary = "^2.9.11"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.4.2"
pytest-flask = "^1.3.0"
pytest-mock = "^3.15.1"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"